requirements.txt
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.12
investpy>=1.0.8
scikit-learn>=1.1.0
statsmodels>=0.13.0
prophet>=1.1.0
pyportfolioopt>=1.5.0
rich>=12.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
sqlalchemy>=1.4.0
pytest>=7.0.0
sphinx>=5.0.0
ollama>=0.1.0
pyyaml>=6.0
